So now we handle the problem of java.lang.System differntly. Previously, we generated some facts to allocate heap objects and then make the library point to them. However this is not extensible, since for me to do that I had to check the c code for the java.lang.System and see how the static fields are initialized. This can change in the future which will render our analysis useless.
The solution to this was to remove all that code for fact generation, and make special objects for all the library classes referenced in the application (I chose the format of "library <class name>"). Then initialize the LibraryPointsTo and make it point to all those special heap objects. This will cause any load of a library static field to assign those special heap objects (along with any other objects the library points to) to the target of the assign. This way, if something in the library changes any of those static fields, we'll still catch it.