modified the LibraryCallGraphEdge for virtual invokes to be as follows: LibraryCallGraphEdge(?inmethod, ?invocation, ?tomethod) :- 	Reachable(?inmethod), 	LibraryMethod(?tomethod), 	VirtualMethodInvocation(?base, ?invocation, ?tomethod, ?inmethod), 	VarPointsTo(?heap, ?base), 	HeapAllocationType(?heap, ?heaptype), 	MethodSignatureType(?tomethod, ?type), 	AssignCompatible(?type, ?heaptype).
This just says that we need to ensure the base of the invocation points somewhere in the heap where the heaptype is compatible with the type of the declaring class of the ?tomethod. No need to actually resolve the method or do a static lookup coz we don't care which method it goes to.