This commit is right before I start solving the bug I found (see below).
I finally found the bug! I'm figuring out a way to solve it now but here's a brief explanation of why this happens (and probably this is behind many of the missing edges too).

So Spark has to operate on whole-program mode to get the call graph. Therefore, it fetches all the the classes used or referenced by the application classes. This of course includes our crafted JAR archive. However, this also includes another set of library classes (namely, the ones that are resolved at a level less than SIGNATURES in the NON-whole-program mode). So the code for the methods of those classes is the original library code which doesn't of course has any reference (e.g., in return statements) to the library points-to set. So eventually, those methods don't return the objects from the LPT as they are supposed to simply because the code that does that is not there to begin with.

The solution I'm developing now will probably involve something along the lines of making our analysis operate on the whole-program too. I'll then just create Jimple bodies for all the concrete methods. This will probably make the size of the resulting JAR file bigger and might (or not) affect the speed of the analysis using it. However, we'll be sacrificing speed for the sake of soundness which is not a bad thing.