commit before solving the return variable bug
I tried sending you the files but for some reason they probably didn't get sent to you. Anyways, I found out the problem. In the JarFactory, I use only one variable of type Object to assign the return of all the method invokes to then assign that variable to the LPT. For some reason, when Soot reads the class file, it generates a new local for the assignment statement of each invocation statement of the same type as the return type of this method. So it doesn't use the variable I create at the very beginning of the method. Consequently, those variables that Soot adds on its own are removed when the phase "jb.dae" is applied. And of course the variable I create is also removed since it's never used in any of those assignment statements.