support for checked exceptions
For checked exception, we now get the types of the exceptions declared
in the catch blocks, create objects of them, call their default
constructors, and their static inialziers (if any).

A problem that remains though is that the guards we currently have are
not enough. They render code after them dead, that's in the "expected"
model we build by hand. So the Java compiler eliminates that code from
the generated class files and they won't show up in the Jimple code we
generate for comparisons.