a more robust guard statement that won't be elimintated by the compiler or Soot via dead code elimination.
I notice that in a few test cases, Eclipse shows some code as dead. For
example, the second guarded throw statement in this method in the
LL.LLSpliterator class

public void forEachRemaining(Consumer action) { 
        	RTA.set = action;
        	if (true) throw new NullPointerException();
        	if (true) throw new ConcurrentModificationException(); 
        	getEst(); // redundant?
        	((Consumer)RTA.set).accept(RTA.set);
        }


The issue now is that the Java compiler eliminates those statements as
well since they are dead code. So they won’t show up in the “expected”
RTA model as they should. This will show as a discrepancy between the
model generated by the RTA analysis and the expected model, while it’s
not. This will be more of an issue when we start to automate the
comparisons.

I've fixed that by adding a boolean field to the RTA class and using
that boolean guard in the condition of the if-statement instead of the
previous simple check 1 != 1