special handling of calls to super or overloaded constructors
if the first non-identity statement is a call to the super constructor
or to another overloaded constructor, we should preserve that call and
keep it the first statement in the generated model. Otherwise, averroes
will generate the default call to the default constructor, which is
wrong. This correctly handles the constructors in the LinkeList test
case.