fixing issues related to handling inner classes
Wow! That was one nasty issue. So for non-static inner classes, in the
constructor, the first statement should be the assignment to the
"this$0" field from the first parameter of the method. If that's the
case, Averroes shouldn't create an assignment to that special field, nor
should it create an assignment of that method parameter.