fixed error related to having null arguments to calls
If the argument to a method call is null, we shouldn't cast set_m to
null, but rather get the declared type of the corresponding method
parameter and cast set_m to that type. This forced us to use the old
for-loop instead of the nice stream API, but oh well!