now XTA generates verifiable bytecode for our example
I had to override the guards for loading the set_f in a method because
it will then be assigned to a set_m with a condition in between. This
construct breaks the flow as the loaded field may not be initialized due
ot the guard. The other solution was to put both the load and the
assignment in the same guard, but that requires quite a bit of change to
XTA, so I decided to go for the easy solution.