fixed unverifiable bytecode due to calls to the overloaded constructor
The call to the overloaded/super constructor is the first statement in
the body of a constructor. Therefore, we cannot depend on casts of the
appropriate set to get the arguments to that call. This fix grabs those
arguments form the arguments to the underlying constructor method that
we are creating a Jimple body for (just like the original code does).