fixed a bug with creating default constructors
The bug is related to calling a non-existent default constructor in the
superclass. This bug took place because of the way classes are
processed; they are not ordered according to hierarchy. I tried
processing them through a TreeSet or a SortedSet but that didn't work.
So I resolved this by inserting a recursive call to create the default
constructor in the direct superclass, if it doesn't already have one.