support for invokedynamic + fixing prepareArguments
I've added support for generating models for method bodies that have
invokedynamic instructions. The way we handle it is keeping the
reference to the bootstrap method as is, and pass a cast as argument,
keeping everything else the same.

The other bug fix addresses an issue when an argument to a specialinvoke
is a primitive type and averroes tries to look for a local variable to
pass as an argument. For that kind of parameters, we'll just send a
default primitive values, just like in regular method calls.