fixed assigning "this" for non-static inner classes
The "this" parameter for non-static inner classes should be assigned to
the underlying set only for non-static methods. In the JVM, unlike our
tests, there are cases where there's a static method (e.g., static
initializer) in a non-static inner class. So we had to restrict this
assignment to avoid a runtime error.