ignore some calls when comparing JSON files
All method calls are from the default constructor of inner classes and to the default constructor of the superclass. This is a hackish way of ignoring such calls, but necessary because Averroes needs those default constructors to generate valid bytecode that involves casting the LPT to the types required to prepare arguments for methods calls to constructors. Without those default constructors, Averroes will just drop dead.