conver any use of SootField to SootFieldRef
Using SootFields forced Averrors to call SootFieldRef.resolve() which we shouldn't. Such a call leads to creating field references that are not necessarily valid according to BCEL.

For example: Soot resolves the field refernece for sun.net.www.URLConnection.url  to java.net.URLConnection.url. Since this field is protected, an attempt to write to that field with the resolved reference in the class sun.net.www.URLConnection leads to a constraint violation in BCEL. So, let's not do it!